requirements.txt
huggingface_hub
streamlit
google-generativeai
python-dotenv
transformers
tensorflow
torchvision
torchaudio